requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
PyYAML==6.0.2
pydantic==2.10.6
python-multipart==0.0.20
pytest==8.3.5
